+++ jest.config.js
@@ -1,8 +1,7 @@
 module.exports = {
     verbose: true,
-    projects:
-    [
-            "<rootDir>/packages/*/jest.config.js"
+    projects: [
+        "<rootDir>/packages/*/jest.config.js"
     ],
     moduleNameMapper: {
         "@qoopido/(.+)$": "<rootDir>/packages/$1/src",
@@ -14,6 +13,9 @@ module.exports = {
     collectCoverageFrom: [
         `<rootDir>/packages/*/src/**/*.js`,
         `!<rootDir>/packages/*/src/**/*.test.js`
+    ],
+    setupFiles: [
+        `<rootDir>/jest.setup.js`
     ],
     coverageDirectory: "<rootDir>/coverage/",
     coverageThreshold: {