+++ packages/validator/jest.setup.js
@@ -1,0 +1,5 @@
+import "jest-canvas-mock";
+import { MessageChannel, MessagePort } from "worker_threads";
+
+self.MessageChannel = MessageChannel;
+self.MessagePort = MessagePort;