+++ packages/pledge/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/dlueth/qoopido/compare/@qoopido/pledge@1.0.10...@qoopido/pledge@1.1.0) (2020-06-19)
+
+
+### Features
+
+* **validator:** add `isTransferable` validation ([6ec0bf9](https://github.com/dlueth/qoopido/commit/6ec0bf9d9966bf042cee4c977d4517399d1671b6))
+
+
+
+
+
 ## [1.0.10](https://github.com/dlueth/qoopido/compare/@qoopido/pledge@1.0.9...@qoopido/pledge@1.0.10) (2020-06-17)
 
 **Note:** Version bump only for package @qoopido/pledge